Move tripal chado Kernel tests into correct dir structure.
